+++ packages/badai-api-example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @badaitech/badai-api-example
 
+## 0.6.11
+
+### Patch Changes
+
+- fix: backward-compatible user ID resolution
+- Updated dependencies
+  - @badaitech/badai-api@0.6.11
+
 ## 0.6.10
 
 ### Patch Changes